+++ build.py
@@ -5,6 +5,6 @@ def build(setup_kwargs):
     setup_kwargs.update(
         {
             "package_data": {"pfun": ["py.typed"]},
-            "ext_modules": cythonize(["pfun/free_effect.pyx"])
+            "ext_modules": cythonize(["pfun/effect.pyx"])
         }
     )
