+++ pfun/logging.py
@@ -1,3 +1,5 @@
+from __future__ import annotations
+
 import logging
 from types import TracebackType
 from typing import Optional, Tuple, Type, Union
